chore(common): Extract `(Owned)LinkedChunkId` into its own module.
This patch extracts the `LinkedChunkId` and `OwnedLinkedChunkId` types
from the `linked_chunk/mod.rs` module to `linked_chunk/identifiers.rs`.

This reduces the `linked_chunk` module by ≈150 LOC.
